Rename CLI Option namespace to Input
