+++ module/CLI/src/Command/Visit/AbstractVisitsListCommand.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\CLI\Command\Visit;
 
-use Shlinkio\Shlink\CLI\Option\EndDateOption;
-use Shlinkio\Shlink\CLI\Option\StartDateOption;
+use Shlinkio\Shlink\CLI\Input\EndDateOption;
+use Shlinkio\Shlink\CLI\Input\StartDateOption;
 use Shlinkio\Shlink\CLI\Util\ExitCodes;
 use Shlinkio\Shlink\CLI\Util\ShlinkTable;
 use Shlinkio\Shlink\Common\Paginator\Paginator;
