+++ module/CLI/src/Input/StartDateOption.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\CLI\Option;
+namespace Shlinkio\Shlink\CLI\Input;
 
 use Cake\Chronos\Chronos;
 use Symfony\Component\Console\Command\Command;
@@ -11,14 +11,14 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 use function sprintf;
 
-class EndDateOption
+class StartDateOption
 {
     private readonly DateOption $dateOption;
 
     public function __construct(Command $command, string $descriptionHint)
     {
-        $this->dateOption = new DateOption($command, 'end-date', 'e', sprintf(
-            'Allows to filter %s, returning only those newer than provided date.',
+        $this->dateOption = new DateOption($command, 'start-date', 's', sprintf(
+            'Allows to filter %s, returning only those older than provided date.',
             $descriptionHint,
         ));
     }
